Merge in MCUCORE/els_pkc from feature/WCBRD-1152-cl-ear5-intake-multiplatform to main

* commit '8bce9e74a8df929628c390dd76b4bb56b5c61679':
  [MCUX-57887] Add cmsis_pack field for mcxn related component
  [MCUX-57887] Add changes needed to els_pkc work on N10
  [MCUX-57887] Add MCXN platform specific components to els_pkc yml
  [WCBRD-1152] replaced usage of 'Els' instead of 'Css" in clib function calls
  [WCBRD-1152] fixed cmsis pack warning about file not available but still being included.
  [WCBRD-1152] -Wint-conversion warning fixed
  [WCBRD-1152] fixed -Wincompatible-pointer-types warnings
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - Yml file added, current supported platform is rw61x.
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - toolchain file renamed
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - toolchain.h file renamed to nxpClToolchain.h (expected new name in next clib-release)
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - nirvana updates re-added in new release from previous release including mcux_els/pkc files new path.
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - psa-driver files updates to have a working build and restore the mcuxClPsaDriver_Oracle_ImportKey function declaration. Both are expected to be fixed in next release of clib.
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - RW61x related mcux_els/pkc files checked in. Include files updated for RW61x (rw612) build
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - rest of files updated
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - 'examples', 'includes', 'src' and 'static_library'  section updated
  [WCBRD-1152] CLNS_SDK_1.2.0 intake - 'doc' section updated

+++ src/comps/mcuxClAeadModes/src/mcuxClAeadModes_ELS_AesGcm.c
@@ -547,7 +547,7 @@ MCUX_CSSL_FP_PROTECTED_TYPE(mcuxClAead_Status_t) mcuxClAead_ModeSkeletonAesGcm(
         /* Call the finalize function. */
         MCUX_CSSL_FP_FUNCTION_CALL(retFinal, pAlgo->pEngine(session, pContext,
                                                                    NULL,
-                                                                   NULL,
+                                                                   (uint32_t)NULL,
                                                                    pContext->partialData,
                                                                    NULL,
                                                                    MCUXCLAEAD_ENGINE_OPTION_FINISH));
@@ -614,7 +614,7 @@ MCUX_CSSL_FP_PROTECTED_TYPE(mcuxClAead_Status_t) mcuxClAead_ModeSkeletonAesGcm(
             MCUX_CSSL_FP_CONDITIONAL((aadProFPFlag1 == 1u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_copy),
                 MCUX_CSSL_FP_CONDITIONAL((aadProFPFlag2 == 1u), pAlgo->protection_token_engine)
             ),
-            MCUX_CSSL_FP_CONDITIONAL((0 != ((adataLength / MCUXCLELS_AEAD_AAD_BLOCK_SIZE) * MCUXCLELS_AEAD_AAD_BLOCK_SIZE)), pAlgo->protection_token_engine),
+            MCUX_CSSL_FP_CONDITIONAL((0u != ((adataLength / MCUXCLELS_AEAD_AAD_BLOCK_SIZE) * MCUXCLELS_AEAD_AAD_BLOCK_SIZE)), pAlgo->protection_token_engine),
             MCUX_CSSL_FP_CONDITIONAL((bytesRemainingAad != 0u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_copy)),
             MCUX_CSSL_FP_CONDITIONAL((aadProFPFlag != 0u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_set),
                                                           pAlgo->protection_token_engine)
@@ -623,7 +623,7 @@ MCUX_CSSL_FP_PROTECTED_TYPE(mcuxClAead_Status_t) mcuxClAead_ModeSkeletonAesGcm(
             MCUX_CSSL_FP_CONDITIONAL((dataProFPFlag1 == 1u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_copy),
                 MCUX_CSSL_FP_CONDITIONAL((dataProFPFlag2 == 1u), pAlgo->protection_token_engine)
             ),
-            MCUX_CSSL_FP_CONDITIONAL((0 != ((inLength / MCUXCLAES_BLOCK_SIZE) * MCUXCLAES_BLOCK_SIZE)), pAlgo->protection_token_engine),
+            MCUX_CSSL_FP_CONDITIONAL((0u != ((inLength / MCUXCLAES_BLOCK_SIZE) * MCUXCLAES_BLOCK_SIZE)), pAlgo->protection_token_engine),
             MCUX_CSSL_FP_CONDITIONAL((bytesRemainingData != 0u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_copy)),
             MCUX_CSSL_FP_CONDITIONAL((dataProFPFlag != 0u), MCUX_CSSL_FP_FUNCTION_CALLED(mcuxClMemory_set),
                                                            pAlgo->protection_token_engine)
